requirements.txt
argcomplete==1.12.3
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
commitizen==2.24.0
decli==0.5.2
distlib==0.3.6
filelock==3.12.2
Flask==2.2.5
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
identify==2.5.24
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
oauthlib==3.2.2
packaging==21.3
platformdirs==3.8.1
pre-commit==2.18.1
prompt-toolkit==3.0.39
pyparsing==3.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
questionary==1.10.0
requests==2.31.0
requests-oauthlib==1.3.1
SQLAlchemy==2.0.18
termcolor==1.1.0
toml==0.10.2
tomlkit==0.11.8
typing_extensions==4.7.1
urllib3==2.0.3
URLObject==2.4.3
virtualenv==20.23.1
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.0
